mostly change in viewmore; customer(add,edit),dashboard,forgotpass remain

+++ src/app/book-page/book-page.component.ts
@@ -15,7 +15,7 @@ import { ViewmorebookComponent } from './viewmorebook/viewmorebook.component';
 })
 export class BookPageComponent implements OnInit {
 
-  displayedColumns: string[] = ['checkbox', 'book_name','product_id', 'product_price', 'product_qty', 'book_description', 'actions'];
+  displayedColumns: string[] = ['checkbox', 'book_name', 'product_id', 'product_price', 'product_qty', 'actions'];
   dataSource: MatTableDataSource<bookClass>;
   deleteFlag: boolean = false;
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
@@ -26,6 +26,23 @@ export class BookPageComponent implements OnInit {
   }
   bookArr: bookClass[] = [];
 
+  ngOnInit() {
+    this._book.getAllBook().subscribe(
+      (data: bookClass[]) => {
+        this.bookArr = data;
+        this.dataSource.data = this.bookArr;
+        this.dataSource.paginator = this.paginator;
+        this.dataSource.sort = this.sort;
+      }
+    );
+  }
+  applyFilter(filterValue: string) {
+    this.dataSource.filter = filterValue.trim().toLowerCase();
+
+    if (this.dataSource.paginator) {
+      this.dataSource.paginator.firstPage();
+    }
+  }
 
   openDialog(row) {
     this._dialog.open(ViewmorebookComponent, {
@@ -44,74 +61,46 @@ export class BookPageComponent implements OnInit {
       this.del_arr.push(item);
       this.deleteFlag = true;
     }
-    console.log(this.del_arr);
   }
+
   onProductListClick() {
     this._route.navigate(['/nav/productList']);
   }
+
   del_arr: bookClass[] = [];
+
   onDeleteSelectedClick() {
     this._book.deleteAllBookData(this.del_arr).subscribe(
       (data: any) => {
-
-        console.log(this.bookArr);
-        console.log(this.del_arr);
-
         for (let i = 0; i < this.del_arr.length; i++) {
           if (this.bookArr.find(x => x == this.del_arr[i])) {
-            this.bookArr.splice(this.bookArr.indexOf(this.del_arr[i]),1);
+            this.bookArr.splice(this.bookArr.indexOf(this.del_arr[i]), 1);
             this.dataSource.data = this.bookArr;
             this.dataSource.paginator = this.paginator;
             this.dataSource.sort = this.sort;
           }
-          console.log(this.del_arr);
-          console.log(this.bookArr);
-
-
         }
-
       });
-
-
-  }
-
-  ngOnInit() {
-    this._book.getAllBook().subscribe(
-      (data: bookClass[]) => {
-        this.bookArr = data;
-        this.dataSource.data = this.bookArr;
-        this.dataSource.paginator = this.paginator;
-        this.dataSource.sort = this.sort;
-        // console.log(data)
-      }
-    );
-
-  }
-  applyFilter(filterValue: string) {
-    this.dataSource.filter = filterValue.trim().toLowerCase();
-
-    if (this.dataSource.paginator) {
-      this.dataSource.paginator.firstPage();
-    }
   }
 
   onAddClick() {
     this._route.navigate(['/nav/AddBook']);
   }
+
   openEdit(row) {
     this._route.navigate(['/nav/EditBook/', row.product_id]);
   }
 
   onDelete(item) {
-    let x: number = this.bookArr.indexOf(item);
-    this._book.deleteBookById(item.product_id).subscribe(
-      (data: any) => {
-        console.log(data)
-        this.bookArr.splice(x, 1);
-        this.dataSource.data = this.bookArr;
-      }
-    );
-    this._route.navigate(['/nav/books']);
+    if (confirm("do you want to delete?")) {
+      let x: number = this.bookArr.indexOf(item);
+      this._book.deleteBookById(item.product_id).subscribe(
+        (data: any) => {
+          this.bookArr.splice(x, 1);
+          this.dataSource.data = this.bookArr;
+        }
+      );
+      this._route.navigate(['/nav/books']);
+    }
   }
-
 }
